fix(backend-async-runner): wire outbox startup and live gates (BUG-0085)
- Attach the production API runner kernel to the outbox dispatcher.
- Fix live upload/privacy gates for MinIO headers, JSONB parameter typing, and repeatable audit tombstones.
- Sync async-runner docs, BUG-0085, retrospective, and work journal evidence.

+++ backend/internal/privacy/runner/store.go
@@ -56,7 +56,7 @@ update privacy_requests
 set status = 'completed',
     completed_at = $2,
     error_code = null,
-    metadata = metadata || jsonb_build_object('deletedFileCount', $3)
+    metadata = metadata || jsonb_build_object('deletedFileCount', $3::int)
 where id = $1 and request_type = 'delete'`, privacyRequestID, now, deletedFileCount)
 }
